Moved sail-related code to dedicated action
