+++ libs/mission/CMakeLists.txt
@@ -13,6 +13,7 @@ add_library(${NAME} STATIC ${SOURCES})
 
 target_link_libraries(${NAME}
     base
+    eps
     time
 )
